Update NTP views naming
