+++ iOS/DuckDuckGo/NewTabPageViewController.swift
@@ -79,14 +79,14 @@ final class NewTabPageViewController: UIHostingController<AnyView>, NewTabPage {
                                                 navigator: DefaultMessageNavigator(delegate: messageNavigationDelegate))
 
         if isNewTabPageCustomizationEnabled {
-            super.init(rootView: AnyView(NewTabPageView(viewModel: self.newTabPageViewModel,
+            super.init(rootView: AnyView(CustomizableNewTabPageView(viewModel: self.newTabPageViewModel,
                                                         messagesModel: self.messagesModel,
                                                         favoritesViewModel: self.favoritesModel,
                                                         shortcutsModel: self.shortcutsModel,
                                                         shortcutsSettingsModel: self.shortcutsSettingsModel,
                                                         sectionsSettingsModel: self.sectionsSettingsModel)))
         } else {
-            super.init(rootView: AnyView(SimpleNewTabPageView(viewModel: self.newTabPageViewModel,
+            super.init(rootView: AnyView(NewTabPageView(viewModel: self.newTabPageViewModel,
                                                               messagesModel: self.messagesModel,
                                                               favoritesViewModel: self.favoritesModel)))
         }
